+++ src/common/io.ts
@@ -50,6 +50,7 @@ import {
   AoEpochSettings,
   AoGateway,
   AoGatewayDelegateWithAddress,
+  AoGatewayVault,
   AoIORead,
   AoIOWrite,
   AoRegistrationFees,
@@ -347,15 +348,8 @@ export class IOReadable implements AoIORead {
         {
           name: 'Timestamp',
           value: (
-            await this.arweave.blocks
-              .getCurrent()
-              .then((block) => {
-                return { timestamp: block.timestamp * 1000 };
-              })
-              .catch(() => {
-                return { timestamp: Date.now() }; // fallback to current time
-              })
-          ).timestamp.toString(),
+            await getCurrentBlockUnixTimestampMs(this.arweave)
+          ).toString(),
         },
       ],
     });
@@ -582,7 +576,9 @@ export class IOReadable implements AoIORead {
       { name: 'Name', value: name },
       {
         name: 'Timestamp',
-        value: timestamp?.toString() ?? Date.now().toString(),
+        value:
+          timestamp?.toString() ??
+          (await getCurrentBlockUnixTimestampMs(this.arweave)).toString(),
       },
       { name: 'Purchase-Type', value: type ?? 'lease' },
       {
@@ -616,6 +612,30 @@ export class IOReadable implements AoIORead {
 
     return this.process.read<PaginationResult<AoDelegation>>({
       tags: pruneTags(allTags),
+    });
+  }
+
+  async getAllowedDelegates(
+    params: PaginationParams & { address: WalletAddress },
+  ): Promise<PaginationResult<WalletAddress>> {
+    return this.process.read<PaginationResult<WalletAddress>>({
+      tags: [
+        { name: 'Action', value: 'Paginated-Allowed-Delegates' },
+        { name: 'Address', value: params.address },
+        ...paginationParamsToTags(params),
+      ],
+    });
+  }
+
+  async getGatewayVaults(
+    params: PaginationParams<AoGatewayVault> & { address: WalletAddress },
+  ): Promise<PaginationResult<AoGatewayVault>> {
+    return this.process.read<PaginationResult<AoGatewayVault>>({
+      tags: [
+        { name: 'Action', value: 'Paginated-Gateway-Vaults' },
+        { name: 'Address', value: params.address },
+        ...paginationParamsToTags(params),
+      ],
     });
   }