+++ tests/e2e/esm/index.test.js
@@ -269,6 +269,50 @@ describe('e2e esm tests', async () => {
       });
     });
 
+    it('should be able get list of allowed gateway delegate addresses, if applicable', async () => {
+      const allowedDelegates = await io.getAllowedDelegates({
+        address: 'QGWqtJdLLgm2ehFWiiPzMaoFLD50CnGuzZIPEdoDRGQ',
+      });
+      assert.ok(allowedDelegates);
+      assert(allowedDelegates.limit === 100);
+      assert(typeof allowedDelegates.totalItems === 'number');
+      assert(typeof allowedDelegates.limit === 'number');
+      assert(typeof allowedDelegates.hasMore === 'boolean');
+      if (allowedDelegates.nextCursor) {
+        assert(typeof allowedDelegates.nextCursor === 'string');
+      }
+      assert(Array.isArray(allowedDelegates.items));
+      allowedDelegates.items.forEach((address) => {
+        assert(typeof address === 'string');
+      });
+    });
+
+    it('should be able to get gateway vaults', async () => {
+      const vaults = await io.getGatewayVaults({
+        address: 'QGWqtJdLLgm2ehFWiiPzMaoFLD50CnGuzZIPEdoDRGQ',
+      });
+      assert.ok(vaults);
+      assert(vaults.limit === 100);
+      assert(vaults.sortOrder === 'desc');
+      assert(vaults.sortBy === 'endTimestamp');
+      assert(typeof vaults.totalItems === 'number');
+      assert(typeof vaults.sortBy === 'string');
+      assert(typeof vaults.sortOrder === 'string');
+      assert(typeof vaults.limit === 'number');
+      assert(typeof vaults.hasMore === 'boolean');
+      if (vaults.nextCursor) {
+        assert(typeof vaults.nextCursor === 'string');
+      }
+      assert(Array.isArray(vaults.items));
+      vaults.items.forEach((vault) => {
+        assert(typeof vault.balance === 'number');
+        assert(typeof vault.cursorId === 'string');
+        assert(typeof vault.vaultId === 'string');
+        assert(typeof vault.startTimestamp === 'number');
+        assert(typeof vault.endTimestamp === 'number');
+      });
+    });
+
     it('should be able to get gateway delegate allow list', async () => {
       const allowList = await io.getGatewayDelegateAllowList({
         address: 'QGWqtJdLLgm2ehFWiiPzMaoFLD50CnGuzZIPEdoDRGQ',